misc(Poll): make spoll the only poll
An advanced poll will be available in the future. This command is *way* too advanced for end-users:

- It has always brought issues and bugs.
- Nobody can figure out for the love of god how to use this, not even me.
- People prefer spoll's behaviour, let's keep things simple!
