+++ src/lib/types/Languages.d.ts
@@ -940,8 +940,6 @@ export interface LanguageKeys {
 		UNKNOWN_USER_SCORE: string;
 		NO_GENRES: string;
 	};
-	COMMAND_POLL_DESCRIPTION: string;
-	COMMAND_POLL_EXTENDED: string;
 	COMMAND_PRICE_DESCRIPTION: string;
 	COMMAND_PRICE_EXTENDED: string;
 	COMMAND_QUOTE_DESCRIPTION: string;
@@ -950,8 +948,8 @@ export interface LanguageKeys {
 	COMMAND_ROLES_EXTENDED: string;
 	COMMAND_SEARCH_DESCRIPTION: string;
 	COMMAND_SEARCH_EXTENDED: string;
-	COMMAND_SPOLL_DESCRIPTION: string;
-	COMMAND_SPOLL_EXTENDED: string;
+	COMMAND_POLL_DESCRIPTION: string;
+	COMMAND_POLL_EXTENDED: string;
 	COMMAND_TOPINVITES_DESCRIPTION: string;
 	COMMAND_TOPINVITES_EXTENDED: string;
 	COMMAND_TOPINVITES_NO_INVITES: string;
@@ -1547,22 +1545,6 @@ export interface LanguageKeys {
 	COMMAND_EMOTES_DESCRIPTION: string;
 	COMMAND_EMOTES_EXTENDED: string;
 	COMMAND_EMOTES_TITLE: string;
-	COMMAND_POLL_MISSING_TITLE: string;
-	COMMAND_POLL_TIME: string;
-	COMMAND_POLL_WANT_USERS: string;
-	COMMAND_POLL_FIRSTUSER: string;
-	COMMAND_POLL_WANT_ROLES: string;
-	COMMAND_POLL_FIRSTROLE: string;
-	COMMAND_POLL_CREATE: (title: string, roles: null | readonly string[], users: null | readonly string[], options: null | readonly string[], time: number, id: string) => string;
-	COMMAND_POLL_LIST_EMPTY: string;
-	COMMAND_POLL_NOTEXISTS: string;
-	COMMAND_POLL_NOTMANAGEABLE: string;
-	COMMAND_POLL_REMOVE: string;
-	COMMAND_POLL_INVALID_OPTION: (options: string) => string;
-	COMMAND_POLL_ALREADY_VOTED: string;
-	COMMAND_POLL_VOTE: string;
-	COMMAND_POLL_MISSING_ID: string;
-	COMMAND_POLL_EMPTY_VOTES: string;
 	COMMAND_PRICE_CURRENCY: (fromCurrency: string, fromAmount: number, toCurrency: string, toAmount: number) => string;
 	COMMAND_PRICE_CURRENCY_NOT_FOUND: string;
 	COMMAND_QUOTE_MESSAGE: string;
