--- src/tasks/pollEnd.ts
@@ -1,9 +1,0 @@
-import { Task } from 'klasa';
-
-export default class extends Task {
-
-	public run() {
-		return Promise.resolve(null);
-	}
-
-}
